show structured response or lg code editor
